refacto: front - move webhook requests cleaning
